feat(nlu-server): model transfer is available without the need of an external server
