+++ packages/nlu-client/src/app-id.ts
@@ -1,0 +1,5 @@
+export const appIdHeader = (appId: string) => {
+  return {
+    'X-App-Id': appId
+  }
+}
