+++ packages/nlu-server/package.json
@@ -37,7 +37,6 @@
     "diff": "^4.0.1",
     "eventemitter2": "^5.0.1",
     "express": "^4.16.4",
-    "express-http-proxy": "1.6.3",
     "express-rate-limit": "^5.5.0",
     "fs-extra": "^9.1.0",
     "glob": "^7.1.6",
@@ -67,7 +66,6 @@
     "@types/cors": "^2.8.12",
     "@types/diff": "^5.0.0",
     "@types/express": "^4.17.13",
-    "@types/express-http-proxy": "1.6.3",
     "@types/express-rate-limit": "^5.1.3",
     "@types/fs-extra": "^5.0.4",
     "@types/glob": "^7.1.3",
